Move avatar fallback into team
