+++ libs/components/src/index.ts
@@ -1,2 +1,1 @@
-export * from "./AvatarFallback";
 export * from "./Typography";
